Changed int to long

+++ week1_programming_challenges/2_max_pairwise_product/MaximumProductPair.java
@@ -4,16 +4,16 @@ public class MaximumProductPair {
 	public static void main(String[] args) {
 		Scanner scanner = new Scanner(System.in);
 		int n = scanner.nextInt();
-		int[] numbers = new int[n];
+		long[] numbers = new long[n];
 		for (int i = 0; i < n; i++) {
-			numbers[i] = scanner.nextInt();
+			numbers[i] = scanner.nextLong();
 		}
 		scanner.close();
 		System.out.println(calculateMaxProduct(numbers));
 	}
 
-	private static int calculateMaxProduct(int[] numbers) {
-		int max = 0, nextMax = 0;
+	private static long calculateMaxProduct(long[] numbers) {
+		long max = 0, nextMax = 0;
 		for (int i = 0; i < numbers.length; i++) {
 			if (numbers[i] > max) {
 				nextMax = max;
